cloudFunctions/read_cloud_storage_ids/requirements.txt
pymongo==3.9.0
google-cloud-storage